chore: 发布v0.1.0正式版，更新所有版本号与文档
- 将所有模块版本从1.0.0修订为0.1.0
- 更新AUTHORS.md、CHANGELOG.md等文档版本信息
- 新增大量工具包、守护进程和测试套件的初始化文件与文档
- 修复部分C语言代码中的内存操作兼容性问题
- 调整CLI网关端点环境变量配置项

+++ agentos/commons/utils/compat/src/compat.c
@@ -75,7 +75,7 @@ int agentos_memset_s(void *dest, int c, size_t dest_size, size_t count)
         return AGENTOS_EINVAL;
     }
 
-    memset(dest, c, count);
+    memset(dest, (int)c, count);
     return 0;
 }
